LayoutTests/inspector-obsolete/runtime/runtime-localStorage-getProperties-expected.txt

Tests RemoteObject.getProperties on localStorage object. 66215

{
    enumerable : false
    name : "testProperty"
    value : {
        description : "testPropertyValue"
        type : "string"
    }
    writable : false
}
